+++ faceshifter/loss.py
@@ -25,25 +25,25 @@ class FaceShifterLoss(LossInterface):
 
         # Attribute loss
         if self.args.W_attr:
-            L_attr = Loss.get_attr_loss(I_t_attr, Y_attr)
+            L_attr = Loss.get_attr_loss(I_t_attr, Y_attr, self.args.batch_size)
             L_G += self.args.W_attr * L_attr
             self.loss_dict["L_attr"] = round(L_attr.item(), 4)
 
         # Reconstruction loss
         if self.args.W_recon:
-            L_recon = Loss.get_L2_loss_with_same_person(Y, I_t, same_person)
+            L_recon = Loss.get_L2_loss_with_same_person(Y, I_t, same_person, self.args.batch_size)
             L_G += self.args.W_recon * L_recon
             self.loss_dict["L_recon"] = round(L_recon.item(), 4)
         
         self.loss_dict["L_G"] = round(L_G.item(), 4)
         return L_G
 
-    def get_loss_D(self, d_true, d_fake):
-        L_true = self.get_hinge_loss(d_true, True)
-        L_fake = self.get_hinge_loss(d_fake, False)
-        L_D = 0.5*(L_true.mean() + L_fake.mean())
+    def get_loss_D(self, d_real, d_fake):
+        L_real = Loss.get_hinge_loss(d_real, True)
+        L_fake = Loss.get_hinge_loss(d_fake, False)
+        L_D = 0.5*(L_real.mean() + L_fake.mean())
         
-        self.loss_dict["L_true"] = round(L_true.mean().item(), 4)
+        self.loss_dict["L_real"] = round(L_real.mean().item(), 4)
         self.loss_dict["L_fake"] = round(L_fake.mean().item(), 4)
         self.loss_dict["L_D"] = round(L_D.item(), 4)
