+++ readme.MD
@@ -1,0 +1,9 @@
+
+
+# test command
+
+python train.py --run_id faceshifter
+python train.py --run_id faceshifter --use_mGPU
+python train.py --run_id faceshifter --use_wandb
+python train.py --run_id faceshifter --use_wandb --use_mGPU
+python train.py --run_id simswap_test --use_wandb --use_mGPU
